net/l2tp: reset skb control buffer
The L2TP stack did not reset the skb control buffer before handing off
the package to the lower netdev.

In a setup with an ath10k radio and batman-adv over an L2TP tunnel
massive fragmentations happen sporadically if the L2TP tunnel is
established over IPv4.

L2TP might reset some of the fields in the IP control buffer, but L2TP
assumes the type of the control buffer to be of an IPv4 packet, while
due to it being Layer 2, this is not a given.

Clear the entire control buffer to avoid such mishaps altogether.

Signed-off-by: David Bauer <mail@david-bauer.net>
(cherry picked from commit 0af1e351787aa2028bd77ed621ff4eb8f5da5e01)