gluon-core: limit size of wireless buffers (#3620)
By default, the kernel configures the buffers for wireless PHYs to 4MB
for 11n radios and 16MB for 11ac and above.

For our RAM constrained devices, this has the effect of the buffers
potentially filling up to cause a OOM oops. This makes it harder to
bisect issues, as non-resumed TX queues often lead to such a crash on
resource-constrained boards.

Limit wireless queue size for devices with 64MB of RAM to 512kB. For
128MB RAM limit to 2MB accordingly. The 512kB was chosen to provide ~100
Mbit/s sustained traffic total to 3 clients on a more powerful board.
This should totally suffice the needs of these lower-tier routers.

Also remove the old logic handling 32MB devices. Gluon dropped support
for those years ago.

Signed-off-by: David Bauer <mail@david-bauer.net>